refactor: rename tui.rs → select_menu.rs, update all callers (#84)
tui.rs was confusingly named (conflicts with tui_app, tui_render, etc).
Renamed to select_menu.rs to clarify it's a standalone crossterm
selection widget, not legacy code.

Updated callers:
- tui_commands.rs: select_menu::select()
- onboarding.rs: select_menu::SelectOption
- commands.rs: select_menu::select()

The select_menu manages its own raw mode internally, so it works
both inside (via raw-mode-disable shim) and outside the TUI viewport.
