+++ koda-cli/src/main.rs
@@ -11,10 +11,10 @@ mod input;
 mod interrupt;
 mod onboarding;
 mod repl;
+mod select_menu;
 mod server;
 mod sink;
 mod tool_history;
-mod tui;
 mod tui_app;
 mod tui_commands;
 mod tui_output;
